up tests for refactoring #222
